Beginning moving of files for better organization going forward
